Cleaned up downloader and events
- Cleaned up playlistdownloader code and internalized a lot of unused methods
- Allowed playlists to be queued with a static method
- Added a lot of documentation with xml comments to events and the downloader
