+++ PlaylistManager/UI/ViewControllers/FoldersViewController.cs
@@ -5,7 +5,7 @@ using BeatSaberMarkupLanguage.FloatingScreen;
 using HMUI;
 using PlaylistManager.HarmonyPatches;
 using PlaylistManager.Interfaces;
-using PlaylistManager.Utilities;
+using PlaylistManager.Downloaders;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
