+++ PlaylistManager/UI/ViewControllers/ImageSelectionModalController.cs
@@ -4,7 +4,7 @@ using BeatSaberMarkupLanguage.Components;
 using BeatSaberMarkupLanguage.Parser;
 using HMUI;
 using PlaylistManager.Types;
-using PlaylistManager.Utilities;
+using PlaylistManager.Downloaders;
 using System;
 using System.Collections.Generic;
 using System.IO;
