+++ PlaylistManager/Utilities/PlaylistLibUtils.cs
@@ -13,7 +13,7 @@ using JetBrains.Annotations;
 using PlaylistManager.Configuration;
 using UnityEngine;
 
-namespace PlaylistManager.Utilities
+namespace PlaylistManager.Downloaders
 {
     public class PlaylistLibUtils
     {
@@ -81,8 +81,14 @@ namespace PlaylistManager.Utilities
             return "";
         }
 
-        public static List<IPlaylistSong> GetMissingSongs(IPlaylist playlist, HashSet<string> ownedHashes = null) => 
-            playlist.Where(s => s.PreviewBeatmapLevel == null && !(ownedHashes?.Contains(s.Hash) ?? false)).Distinct(IPlaylistSongComparer<IPlaylistSong>.Default).ToList();
+        public static List<IPlaylistSong> GetMissingSongs(IPlaylist playlist, HashSet<string> ownedHashes = null)
+        {
+            if (playlist != null)
+            {
+                return playlist.Where(s => s.PreviewBeatmapLevel == null && !(ownedHashes?.Contains(s.Hash) ?? false)).Distinct(IPlaylistSongComparer<IPlaylistSong>.Default).ToList();
+            }
+            return new List<IPlaylistSong>();
+        }
 
         #region Image
